refact: tidy tls utils
